bookkeepr/requirements.txt
# Core Framework
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3

# Database
psycopg2-binary==2.9.9

# QuickBooks Integration
intuit-oauth==1.2.4
python-quickbooks==0.9.5
requests==2.31.0

# Task Queue
celery==5.3.4
redis==5.0.1

# AI / LLM
openai==1.6.1

# Environment Management
python-dotenv==1.0.0

# Security
bcrypt==4.1.2
pyjwt==2.8.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
phonenumbers==8.13.27

# Development
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Production
gunicorn==21.2.0

flask-login>=0.6.3
flask-migrate>=4.0.5
pytest>=7.4.3
pytest-flask>=1.3.0
